fix: moved OrderStore.GiftCardError under Yosemite layer because of warning generated by swift proposal SE-0364
